Clarify Array Slicing Syntax from 'array[start:stop[:step]]' to 'array[start:stop:step]' (#465)
* Clarified Array Slicing Syntax

* Update numpy-basics.ipynb

Thank you for the feedback. I appreciate the concern about the notation implying that the second colon and the step parameter are mandatory. 

I propose we adopt 'array[start:stop:step]' as the primary notation in examples but include a clear note in the text explaining that the step parameter is optional: "The step parameter is optional and can be omitted, in which case the slice uses a default step of 1."

* Resolved issue with wrong use of parenthesis ` `